housekeeping
updated file structure to include a main.js
added comments, rewrote others
renamed scss partials to being with _
fixed sticky navbar to use css instead of javascript
