+++ details/style/_extends.scss
