+++ details/style/_genericMixins.scss
