+++ details/style/_nav.scss
