+++ details/style/_reset.scss
