+++ details/style/_search.scss
@@ -10,6 +10,7 @@
     height: 8rem !important;
     margin-left: auto;
     margin-right: auto;
+
     input[type="text"] {
         @extend %secondary-colors;
         width: 90%;
@@ -17,6 +18,7 @@
         border-radius: 10px;
         margin: 0 auto;
     }
+
     #searchResults {
         @extend %flex-column-nowrap;
         p {
@@ -32,6 +34,7 @@
         margin-top: 2rem;
         width: 100%;
         top: 2rem;
+
         #searchResults {
             p {
                 text-align: center;
