+++ details/style/_sort.scss
