+++ details/style/_variables.scss
