--- details/style/body.scss
@@ -1,6 +1,0 @@
-body {
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    @extend %primary-colors;
-}
