+++ style/home/_footer.scss
