+++ style/home/_spiritBox.scss
